+++ podcasts/MediaFilterOverlayController.swift
@@ -84,10 +84,10 @@ class MediaFilterOverlayController: FilterSettingsOverlayController, UITableView
         cell.style = .primaryUi01
         cell.setSelectState(selectedIndex == indexPath.row)
         if FeatureFlag.playlistsRebranding.enabled {
-            cell.title.font = .systemFont(ofSize: 17, weight: .semibold)
+            cell.title.font = .font(ofSize: 17, weight: .semibold, scalingWith: .body)
             cell.setTintColor(color: AppTheme.colorForStyle(.primaryInteractive01))
         } else {
-            cell.title.font = .systemFont(ofSize: 16, weight: .medium)
+            cell.title.font = .font(ofSize: 16, weight: .medium, scalingWith: .callout)
             cell.setTintColor(color: filterToEdit.playlistColor())
         }
         cell.selectButton.tag = indexPath.row
@@ -101,6 +101,10 @@ class MediaFilterOverlayController: FilterSettingsOverlayController, UITableView
     }
 
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return UITableView.automaticDimension
+    }
+
+    func tableView(_ tableView: UITableView, estimatedHeightForRowAt indexPath: IndexPath) -> CGFloat {
         FeatureFlag.playlistsRebranding.enabled ? 46 : 51
     }