+++ podcasts/New Detail/Playlist Header/PlaylistHeaderView.swift
@@ -4,6 +4,8 @@ struct PlaylistHeaderView: View {
     @EnvironmentObject var theme: Theme
     @ObservedObject var viewModel: PlaylistDetailViewModel
 
+    @ScaledMetric(relativeTo: .largeTitle) private var iconSize = CGFloat(18)
+
     var description: String {
         let duration = viewModel.totalDuration()
         switch viewModel.playlistEpisodesCount {
@@ -91,17 +93,17 @@ struct PlaylistHeaderView: View {
         Button {
             action(type)
         } label: {
-            HStack(alignment: .top, spacing: 8.0) {
+            HStack(alignment: .center, spacing: 8.0) {
                 image
                     .renderingMode(.template)
                     .resizable()
                     .foregroundStyle(color)
                     .scaledToFit()
-                    .frame(width: 18, height: 18)
+                    .frame(width: iconSize, height: iconSize)
                 Text(title)
                     .font(style: .subheadline, weight: .medium)
                     .foregroundStyle(color)
-                    .lineLimit(1)
+                    .lineLimit(2)
                     .multilineTextAlignment(.leading)
                     .fixedSize(horizontal: false, vertical: true)
             }